+++ lib/karafka/web/pro/ui/controllers/consumers/commands_controller.rb
@@ -60,7 +60,7 @@ module Karafka
                 # If nothing found, lets redirect user back to the commands list
                 redirect(
                   'commands',
-                  warning: 'No recent commands available.'
+                  warning: 'No recent commands available'
                 )
               end