+++ lib/karafka/web/pro/ui/controllers/topics/topics_controller.rb
@@ -27,12 +27,37 @@ module Karafka
               # Renders form for creating a new topic with basic details like number of partitions
               # and the replication factor
               def new
-                raise
+                only_with_management_active!
+
+                render
               end
 
               # Creates topic and redirects on success
+              # @note Upon creation we do not redirect to the topic config page because it may take
+              #   a topic a moment to be fully available in the cluster. We "buy" ourselves this
+              #   time by redirecting user back to the topics list
               def create
-                raise
+                only_with_management_active!
+
+                begin
+                  Karafka::Admin.create_topic(
+                    params[:topic_name],
+                    params.int(:partitions_count),
+                    params.int(:replication_factor)
+                  )
+                rescue Rdkafka::RdkafkaError => e
+                  @form_error = e
+                end
+
+                return new if @form_error
+
+                redirect(
+                  'topics',
+                  success: format_flash(
+                    'Topic ? successfully created',
+                    params[:topic_name]
+                  )
+                )
               end
             end
           end