+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **[Feature]** Support consumers that have mismatching schema.
 - **[Feature]** Provide ability to navigate to a timestamp in the Explorer (Pro).
 - **[Feature]** Provide ability to manage topics and their configuration from the Web UI (Pro).
+- [Enhancement] Improve flash messages.
 - [Enhancement] Improve handling of post-submit redirects.
 - [Enhancement] Provide better support for fully transactional consumers.
 - [Enhancement] Error out when `#setup` is called after `#enable!`.